Fixed bug in erase
The bug was occurring when we were releasing the node to_erase before calling transpose_subtree. Doing that implies the function transpose_subtree cannot evaluate anymore the "child type" of the node to_erase  (e.g if to_erase is the left child of its parent, then its child type is "left").
The consequences of this are memory leaks and wrong child/parent assignments.